requirements.txt
fastapi
uvicorn
uuid
piper-tts
python-dotenv
httpx
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
jarvis-settings-client @ git+https://github.com/alexberardi/jarvis-settings-client.git@main
jarvis-auth-client @ git+https://github.com/alexberardi/jarvis-auth-client.git@main
jarvis-config-client @ git+https://github.com/alexberardi/jarvis-config-client.git@main
jarvis-log-client @ git+https://github.com/alexberardi/jarvis-log-client.git@main
